refactor(import): reorganize PDF import components and update localization
- Refactored the import button and type selection components to use new paths for PDF upload and processing.
- Updated the import type from 'customCsv' to 'intelligentImport' in localization files for better clarity.
- Removed obsolete PDF processing and upload components to streamline the import functionality.
- Enhanced the handling of processor components to ensure proper step validation during imports.
