+++ app/[locale]/dashboard/components/import/components/format-preview.tsx
@@ -26,7 +26,7 @@ import {
 } from "@tanstack/react-table";
 import React from "react";
 import { experimental_useObject as useObject } from '@ai-sdk/react'
-import { tradeSchema } from '@/app/api/ai/format-trades/schema'
+import { tradeSchema } from '@/app/api/imports/ibkr/format-trades/schema'
 import { z } from 'zod'
 import { Badge } from "@/components/ui/badge";
 import {
